SummerHacks/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.22.1",
    "eth-account>=0.13.7",
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
    "sse-starlette>=3.3.4",
    "supabase>=2.28.3",
    "uvicorn>=0.44.0",
    "web3>=7.15.0",
]
